Thread safety
A lock around all operations that write to the socket so that writes
can't be intermingled. Shouldn't need a lock for reads as only the
read_loop reads from the socket.

Spawn message_loop in same thread as read_loop
So that processing messages don't require a thread context switch.