Reader and writer separation (#8)
* Thread safety

A lock around all operations that write to the socket so that writes
can't be intermingled. Shouldn't need a lock for reads as only the
read_loop reads from the socket.

Spawn message_loop in same thread as read_loop
So that processing messages don't require a thread context switch.

* Use same thread of socket connection and read_loop/message_loop

* Use trace log level only in hot path, otherwise debug

* Write loop

* pass socket too all send_ methods

* Connection Reader and Writer seperation

* Drop same_thread for EC compablility

* more compact Packets using records